feat: Update Project model to use UUID for id field
The Project model in the agentic_preview application has been updated to use UUID for the id field instead of a regular string. This change ensures uniqueness and improves the security of the identifiers. The id field now has a default value generated using the uuid4 function from the uuid module.

# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
# On branch main
# Your branch is up to date with 'origin/main'.
#
# Changes to be committed:
#       modified:   agentic_preview/agentic_preview/models.py
#